Introduced (potentially) reusable `RetryableDeclarativeAgent`.
