+++ pipeline-model-definition/src/main/java/org/jenkinsci/plugins/pipeline/modeldefinition/agent/impl/Label.java
@@ -29,7 +29,6 @@ import hudson.Util;
 import hudson.util.FormValidation;
 import org.apache.commons.lang.StringUtils;
 import org.jenkinsci.Symbol;
-import org.jenkinsci.plugins.pipeline.modeldefinition.agent.DeclarativeAgent;
 import org.jenkinsci.plugins.pipeline.modeldefinition.agent.DeclarativeAgentDescriptor;
 import org.kohsuke.stapler.DataBoundConstructor;
 import org.kohsuke.stapler.DataBoundSetter;
@@ -38,11 +37,11 @@ import org.kohsuke.stapler.QueryParameter;
 import edu.umd.cs.findbugs.annotations.CheckForNull;
 import edu.umd.cs.findbugs.annotations.NonNull;
 import edu.umd.cs.findbugs.annotations.Nullable;
+import org.jenkinsci.plugins.pipeline.modeldefinition.agent.RetryableDeclarativeAgent;
 
-public class Label extends DeclarativeAgent<Label> {
+public class Label extends RetryableDeclarativeAgent<Label> {
     private String label;
     private String customWorkspace;
-    private int retries = 1;
 
     @DataBoundConstructor
     public Label(String label) {
@@ -64,16 +63,7 @@ public class Label extends DeclarativeAgent<Label> {
         this.customWorkspace = customWorkspace;
     }
 
-    public int getRetries() {
-        return retries;
-    }
-
-    @DataBoundSetter
-    public void setRetries(int retries) {
-        this.retries = retries;
-    }
-
-    @Extension(ordinal = -800) @Symbol({"label","node"})
+    @Extension(ordinal = -800) @Symbol({"label","node"}) // TODO perhaps put node first, for AgentDirective
     public static class DescriptorImpl extends DeclarativeAgentDescriptor<Label> {
         @Override
         @NonNull
