+++ pipeline-model-extensions/src/main/resources/org/jenkinsci/plugins/pipeline/modeldefinition/agent/RetryableDeclarativeAgent/help-retries.html
@@ -1,0 +1,3 @@
+<div>
+    If set to a number greater than one, some agent- or restart-related failures can be recoverable.
+</div>
